Checking if token works on ec2
